fix: stop swallowing failures, and give the raises a type
Second Phase 1 batch. Findings 341 -> 135, but three of these are real
defects rather than lint cleanup.

**GDPR erasure claimed completion it had not achieved.**
`anonymise_and_delete_user` wraps the authenticator and session deletes
in `except Exception: pass`, then logs "GDPR deletion complete". Its own
docstring promises "a failure halfway through leaves the user intact",
so swallowing contradicted the documented design: a failed DELETE left
those records in place while the log asserted erasure. Only `ImportError`
is tolerated now — that genuinely means the optional allauth app is not
installed — and an operational failure propagates, aborting the
transaction exactly as the docstring says it should. The same shape in
the export path silently dropped a category of personal data from a
right-of-access response that still presented itself as complete.

**The retention clock read the server's date.** `latest_invoice_year`
fell back to `date.today().year`. UTC lags Europe/Athens, so around New
Year that returns the EARLIER year — shortening a legal retention window,
which that function's docstring calls the unrecoverable direction of the
two possible mistakes. Now `timezone.localdate().year`, with the test
following rather than pinning the old behaviour. The rest of DTZ went the
same way: `USE_TZ` is True, so Django's `timezone.now()` is the correct
call, not ruff's suggested `datetime.now(tz=...)`.

**`meili` raised bare `Exception` on every failed task**, and
`meili/apps.py` raised `Exception` then caught `Exception` one line
later — so a genuine `TypeError` in that block was indistinguishable from
an indexing failure. New `meili/exceptions.py` and `core/exceptions.py`
follow the per-app convention `order`/`shipping*`/`meta_capi` already
use. The core ones matter because those tasks declare
`autoretry_for=(Exception,)`: "the monitored dependency is down" is worth
retrying, "this task has a bug" is not, and there was no way to tell them
apart.

Also narrowed `tenant/credentials.py` (its comment always claimed to
guard a missing optional app; as written it also hid a database failure,
and that function decides the From address on every outbound email) and
made the platform dashboard say so when a tenant's revenue cannot be
read, instead of reporting 0.00 — indistinguishable from a store that
took no money.

RUF046 removed `int()` around `round(Decimal)` in the Stripe and Viva
cents conversion. Verified equivalent across bankers-rounding boundaries
(0.005, 2.5, 3.5, 19.995, negatives) before applying: `round(Decimal)`
already returns an int.

Test exemptions are measured, not assumed — TRY002, DJ008 and S110 are
100% test-only, and each conftest swallow already carried a comment
saying it emulates `on_commit(robust=True)`.

+++ admin/dashboard.py
@@ -938,7 +938,7 @@ def _zone_c_queues(Order, ProductReview, Contact, ReviewStatus) -> dict:
                 "user_label": (
                     review.user.email if review.user else str(_("Anonymous"))
                 )[:25],
-                "rate": max(0, min(10, int(round(float(review.rate or 0))))),
+                "rate": max(0, min(10, round(float(review.rate or 0)))),
                 "created_at": review.created_at,
             }
         )
